Foreign.NVVM.Error

Description

Error handling

Documentation

data Status

NVVM API function return code

Constructors

Success

OutOfMemory

ProgramCreationFailure

IRVersionMismatch

InvalidInput

InvalidProgram

InvalidIR

InvalidOption

NoModuleInProgram

CompilationFailure

describe :: Status -> String

resultIfOk :: (Status, a) -> IO a

Return the result of a function on successful execution, otherwise throw an exception.

nothingIfOk :: Status -> IO ()

Throw an exception on an unsuccessful return code

checkStatus :: CInt -> IO ()

nvvmError :: String -> a

Throw an exception. Exceptions may be thrown from pure code, but can only be caught in the IO monad.

nvvmErrorIO :: String -> IO a

Raise an NVVM exception in the IO monad

requireSDK :: Name -> Double -> a

A specially formatted error message