Foreign.NVVM.Compile

Description

Program compilation

Documentation

data Program

An NVVM program

data Result

The result of compiling an NVVM program.

Constructors

Result

Fields

compileResult :: !ByteString

The compiled kernel, which can be loaded into the current program using 'Foreign.CUDA.Driver.loadData*'

compileLog :: !ByteString

Warning messages generated by the compiler/verifier

data CompileOption

Program compilation options

Constructors

OptimisationLevel !Int
optimisation level, from 0 (disable optimisations) to 3 (default)

Target !Compute
target architecture to compile for (default: compute 2.0)

FlushToZero
flush denormal values to zero when performing single-precision floating-point operations (default: no)

NoFMA
disable fused-multiply-add instructions (default: enabled)

FastSqrt
use a fast approximation for single-precision floating-point square root (default: no)

FastDiv
use a fast approximation for single-precision floating-point division and reciprocal (default: no)

GenerateDebugInfo
generate debugging information (-g) (default: no)

compileModule

Arguments

:: String
name of the module

-> ByteString
NVVM IR in either textual or bitcode representation

-> [CompileOption]
compiler options

-> IO Result

Compile an NVVM IR module, in either bitcode or textual representation, into PTX code.

compileModules

Arguments

:: [(String, ByteString)]
(module name, module NVVM IR) pairs to compile

-> [CompileOption]
compiler options

-> IO Result

Compile a collection of NVVM IR modules into PTX code

create :: IO Program

Create an empty Program

http://docs.nvidia.com/cuda/libnvvm-api/group__compilation.html#group__compilation_1g46a0ab04a063cba28bfbb41a1939e3f4

destroy :: Program -> IO ()

Destroy a Program

http://docs.nvidia.com/cuda/libnvvm-api/group__compilation.html#group__compilation_1gfba94cab1224c0152841b80690d366aa

addModule

Arguments

:: Program
NVVM program to add to

-> String
Name of the module (defaults to "<unnamed>" if empty)

-> ByteString
NVVM IR module in either bitcode or textual representation

-> IO ()

Add a module level NVVM IR to a program

http://docs.nvidia.com/cuda/libnvvm-api/group__compilation.html#group__compilation_1g0c22d2b9be033c165bc37b16f3ed75c6

addModuleFromPtr

Arguments

:: Program
NVVM program to add to

-> String
Name of the module (defaults to "<unnamed>" if empty)

-> Int
Number of bytes in the module

-> Ptr Word8
NVVM IR module in bitcode or textual representation

-> IO ()

As with addModule, but read the specified number of bytes from the given pointer.

compile :: Program -> [CompileOption] -> IO (ByteString, Maybe ByteString)

Compile the NVVM program. Returns the log from the compiler/verifier and, if successful, the compiled program.

http://docs.nvidia.com/cuda/libnvvm-api/group__compilation.html#group__compilation_1g76ac1e23f5d0e2240e78be0e63450346

verify :: Program -> [CompileOption] -> IO (Status, ByteString)

Verify the NVVM program. Returns whether compilation will succeed, together with any error or warning messages.